+++ components/schemes/Color/tests/unit/MDCTonalPaletteTests.m
@@ -40,15 +40,15 @@
 
   // Given
   NSArray<UIColor *> *tonalColors = @[
-    [UIColor colorWithRed:(CGFloat)0.6 green:(CGFloat)0.2 blue:(CGFloat)0.3 alpha:1],
-    [UIColor colorWithRed:1 green:(CGFloat)0.5 blue:0 alpha:1],
-    [UIColor colorWithRed:(CGFloat)0.2 green:(CGFloat)0.3 blue:(CGFloat)0.3 alpha:1],
-    [UIColor colorWithRed:(CGFloat)0.5 green:(CGFloat)0.8 blue:(CGFloat)0.6 alpha:1],
-    [UIColor colorWithRed:(CGFloat)0.4 green:(CGFloat)0.7 blue:(CGFloat)0.4 alpha:1],
-    [UIColor colorWithRed:(CGFloat)0.8 green:(CGFloat)0.3 blue:(CGFloat)0.9 alpha:1],
-    [UIColor colorWithRed:(CGFloat)0.6 green:(CGFloat)0.2 blue:(CGFloat)0.3 alpha:1],
-    [UIColor colorWithRed:0 green:(CGFloat)0.8 blue:1 alpha:1],
-    [UIColor colorWithRed:(CGFloat)0.1 green:(CGFloat)0.8 blue:(CGFloat)0.3 alpha:1],
+    [UIColor colorWithRed:(float)0.6 green:(float)0.2 blue:(float)0.3 alpha:1],
+    [UIColor colorWithRed:1 green:(float)0.5 blue:0 alpha:1],
+    [UIColor colorWithRed:(float)0.2 green:(float)0.3 blue:(float)0.3 alpha:1],
+    [UIColor colorWithRed:(float)0.5 green:(float)0.8 blue:(float)0.6 alpha:1],
+    [UIColor colorWithRed:(float)0.4 green:(float)0.7 blue:(float)0.4 alpha:1],
+    [UIColor colorWithRed:(float)0.8 green:(float)0.3 blue:(float)0.9 alpha:1],
+    [UIColor colorWithRed:(float)0.6 green:(float)0.2 blue:(float)0.3 alpha:1],
+    [UIColor colorWithRed:0 green:(float)0.8 blue:1 alpha:1],
+    [UIColor colorWithRed:(float)0.1 green:(float)0.8 blue:(float)0.3 alpha:1],
     [UIColor colorWithRed:1 green:0 blue:1 alpha:1]
   ];
